add `define` export for custom element registration
